SciAgent/requirements.txt
streamlit==1.31.0
langchain==0.1.19
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.0.5
langgraph==0.0.25
pypdf==3.17.1
python-dotenv==1.0.0
tiktoken==0.5.2
pydantic==2.5.3
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers
openai==1.9.0
torch